requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
pydantic==2.1.1
pydantic-core==2.4.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.27.0
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
